fix: move hash function to utils and consolidate
