+++ src/components/PayPalButtons.tsx
@@ -6,7 +6,7 @@ import React, {
     ReactElement,
 } from "react";
 import { usePayPalScriptReducer } from "../ScriptContext";
-import { getPayPalWindowNamespace, DEFAULT_PAYPAL_NAMESPACE } from "./utils";
+import { getPayPalWindowNamespace, DEFAULT_PAYPAL_NAMESPACE } from "../utils";
 import type {
     PayPalButtonsComponentOptions,
     PayPalButtonsComponent,
