+++ src/utils.ts
@@ -8,3 +8,13 @@ export function getPayPalWindowNamespace(
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     return (window as any)[namespace];
 }
+/**
+ * Creates a numeric hash based on the string input.
+ */
+export function hashStr(str: string): number {
+    let hash = 0;
+    for (let i = 0; i < str.length; i++) {
+        hash += str[i].charCodeAt(0) * Math.pow((i % 10) + 1, 5);
+    }
+    return Math.floor(Math.pow(Math.sqrt(hash), 5));
+}
